% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_studies.R
\name{add_corrected_errors}
\alias{add_corrected_errors}
\title{Add Corrected Errors to Screening Decisions}
\usage{
add_corrected_errors(con, syrf_file = "", sys_search_name = "")
}
\arguments{
\item{con}{A database connection object.}

\item{syrf_file}{A character string specifying the file path containing corrected screening decisions.}

\item{sys_search_name}{A character string specifying the name of the systematic search for the re-screened studies.}
}
\value{
This function does not explicitly return a value. It updates the database with corrected
screening decisions if specified and saves pre and post-correction human decisions to .fst files.
}
\description{
Incorporates corrected screening decisions into the database.
This function requires a database connection object, a file containing corrected screening decisions, and the systematic search name.
}
