% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{check_if_retrieved}
\alias{check_if_retrieved}
\title{Update retrieved citations table}
\usage{
check_if_retrieved(con, citations)
}
\arguments{
\item{con}{connection to db}

\item{citations}{df of all new citations retrieved}
}
\value{
update of the retrieved_citations table
}
\description{
This function updates the retrieved citations table with newly retrieved studies
}
