% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{check_na_cols}
\alias{check_na_cols}
\title{Check and Summarize NA Counts in Each Column}
\usage{
check_na_cols(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A list containing data frames, each representing the count of NAs for a specific column.
}
\description{
This function calculates and summarizes the count of NAs for each column in a data frame.
It groups the data by each column and provides the count of NAs within each group.
}
\details{
This function iterates through each column in the data frame and groups the data by the values in that column.
It then calculates the count of NAs for each group and stores the results in a list. The list is printed, displaying
the NA counts for each column.
}
\examples{
\dontrun{
# Example usage:
data <- data.frame(
  source = c("A", "B", "A", "B", "A"),
  col1 = c(1, 2, NA, 4, 5),
  col2 = c(NA, 2, 3, 4, 5)
)
check_na_cols(data)
}

}
