% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pico_utils.R
\name{check_pico}
\alias{check_pico}
\title{This function checks the status of pico tables in a SOLES project and provides users the option of
updating to the newest version available on OSF.}
\usage{
check_pico(con, master_node)
}
\arguments{
\item{con}{Connection to soles project db}

\item{master_node}{OSF node}
}
\description{
This function checks the status of pico tables in a SOLES project and provides users the option of
updating to the newest version available on OSF.
}
\examples{
\dontrun{
# Example usage:
check_pico(con)
}

}
