% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{check_tables}
\alias{check_tables}
\title{Check and optionally delete redundant tables in the database.}
\usage{
check_tables(con, delete_redundant = FALSE)
}
\arguments{
\item{con}{A database connection object.}

\item{delete_redundant}{Logical. If TRUE, delete redundant tables after
saving them as .fst files.}
}
\value{
The function doesn't return anything, but prints messages indicating
whether column names match and actions taken.
}
\description{
This function checks whether the column names of each table in the database
match the expected column names. Optionally, it can delete tables that are
redundant (not expected). It also checks the format of the scopus and wos
uids in the relevant columns, as well as the doi column in each table. Providing
the user with an option to update them to the correct format where necessary.
}
\examples{
\dontrun{
# Connect to the database
con <- pool::dbPool(RPostgres::Postgres(),
                    dbname = Sys.getenv("ad_soles_dbname"),
                    host = Sys.getenv("ad_soles_host"),
                    port = 5432,
                    user = Sys.getenv("ad_soles_user"),
                    password = Sys.getenv("ad_soles_password"))

# Check tables and delete redundant ones
check_tables(con, delete_redundant = FALSE)

# Close the database connection
poolClose(con)
}

}
