% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pico_utils.R
\name{clean_regex_file}
\alias{clean_regex_file}
\title{Go through regex input file, identify duplicate names and regexes and optionally remove them}
\usage{
clean_regex_file(regex_file)
}
\arguments{
\item{regex_file}{The input file containing the regex dictionary. Required column names:
name, type, regex, main_category, sub_category1 and sub_category2}
}
\description{
Go through regex input file, identify duplicate names and regexes and optionally remove them
}
\keyword{internal}
