% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{combine_searches}
\alias{combine_searches}
\title{Combine searches}
\usage{
combine_searches(...)
}
\arguments{
\item{...}{One or more data frame objects to be combined. Each data frame should represent search results
and can have different sets of columns. The function automatically aligns these data frames by column names,
appending missing columns as \code{NA}.}
}
\value{
combined search results with all the correct columns
}
\description{
This function combines search results across searches
}
