% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pico_tag.R
\name{complete_pico_tag}
\alias{complete_pico_tag}
\title{Complete PICO Tagging Function}
\usage{
complete_pico_tag(
  con,
  tag_method = "",
  tag_type = "",
  tag_main_category = "all"
)
}
\arguments{
\item{con}{A database connection object.}

\item{tag_method}{The tagging method to use, either "fulltext" or "tiabkw".}

\item{tag_type}{The type of PICO tag to apply (e.g. "intervention", "species" etc).}

\item{tag_main_category}{The main category of the PICO tag. Default is "all".}
}
\value{
The function does not explicitly return a value, but it updates the connected database with new pico tags.
}
\description{
This function checks how many studies are still to be tagged, then runs the pico_tag function multiple times until there are no more studies to tag.
}
\details{
This function retrieves studies that are already tagged using the specified tagging method, and then tags the remaining studies that meet
the specified criteria. It supports two tagging methods: "fulltext" and "tiabkw". The tagging process
involves using regular expressions to extract relevant information from the full text or title/abstract/keywords, depending on
the chosen method.
}
\examples{
\dontrun{
Tag studies using full text regex
complete_pico_tag(
  con, 
  tag_method = "fulltext", 
  tag_type = "intervention", 
  tag_main_category = "all"
)
}
}
