% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{completionPieServer}
\alias{completionPieServer}
\title{Generate a Pie Chart for Tagging Completion}
\usage{
completionPieServer(
  id,
  table,
  identifier,
  included_studies,
  remove_failed = FALSE
)
}
\arguments{
\item{id}{The module identifier.}

\item{table}{The data table to use for generating the chart.}

\item{identifier}{The name of the column to use for generating the chart.}

\item{included_studies}{The table used for bringing in included studies.}

\item{remove_failed}{Logical, indicating whether to remove rows with status "failed". Default is FALSE.}
}
\description{
This server function generates a pie chart indicating the completion percentage of tagging based on the provided inputs. The chart represents the percentage of occurrences of the values in the specified column for each category (complete or not complete). It also offers an option to remove failed status rows and displays the completion percentage with a color-coded legend.
}
