% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{completionPieUI}
\alias{completionPieUI}
\title{Generate a Pie Chart Indicating Completion}
\usage{
completionPieUI(id)
}
\arguments{
\item{id}{The module identifier.}
}
\description{
This Shiny module creates a pie chart indicating the completion percentage based on a specified table and column. The chart displays the count of occurrences of the values of the selected column and includes a legend with a color-coded representation of those values. Intended for use within a Shiny application.
}
