% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{create_tables}
\alias{create_tables}
\title{This function creates the required tables for the soles database}
\usage{
create_tables(con, overwrite = FALSE)
}
\arguments{
\item{con}{Connection to db}

\item{overwrite}{Whether to overwrite an existing table. Default is \code{FALSE}.}
}
\description{
This function creates the required tables for the soles database
}
\examples{
\dontrun{
create_tables(con)
create_tables(con, overwrite = TRUE)
}

}
