% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup.R
\name{dedup_first_search}
\alias{dedup_first_search}
\title{Get new unique citations and perform deduplication}
\usage{
dedup_first_search(citations, arrange_by = NULL, keep_source = "pubmed")
}
\arguments{
\item{citations}{A dataframe containing citation information.}

\item{arrange_by}{Column in the dataframe to arrange data by, before duplication starts. Choose from "year", "author" or "title".}

\item{keep_source}{source to preferentially retain over other sources}
}
\value{
A list with components:
\describe{
\item{unique}{Dataframe containing unique citations.}
\item{manual_dedup}{Dataframe containing citations to be manually checked for duplicates.}
\item{match_ids}{Dataframe containing updated matching identifiers.}
}
}
\description{
This function processes a dataframe of citations, deduplicates them, and provides unique and manually deduplicated citation results.
}
