% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{download_table_UI}
\alias{download_table_UI}
\title{Download table UI}
\usage{
download_table_UI(id)
}
\arguments{
\item{id}{The module identifier.}
}
\description{
This Shiny module UI function creates the UI
}
