% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ft.R
\name{elsevier_ft}
\alias{elsevier_ft}
\title{Elsevier full text retrieval}
\usage{
elsevier_ft(doi, uid, token, path)
}
\arguments{
\item{doi}{digital object identifier}

\item{uid}{unique identifier for paper}

\item{token}{API token}

\item{path}{where full texts should be stored}
}
\description{
This function retrieves full texts from Elsevier journals
}
