% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_api.R
\name{epmc_search}
\alias{epmc_search}
\title{Search EuropePMC and retrieve bibliographic data using europepmc}
\usage{
epmc_search(query = NULL, timespan, retMax = 5000, format_soles = TRUE)
}
\arguments{
\item{query}{a character string containing a correctly syntaxed EuropePMC search}

\item{timespan}{a formatted character string defining the timespan you want to search}

\item{retMax}{The maximum number of records to retrieve, default is 5000, maximum is 5000}

\item{format_soles}{boolean, if set to TRUE will format search results for the SOLES workflow, default is TRUE}
}
\value{
a dataframe containing EuropePMC search results
}
\description{
A wrapper function for europepmc. Search EuropePMC using a query and retrieve results programmatically.
}
\examples{
\dontrun{
query <- "TITLE_ABS:(dementia OR \"memory loss\")"
epmc_result <- epmc_search(query, timespan, retMax = 500, format_soles = FALSE)
}
}
