% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{evidence_map_Server}
\alias{evidence_map_Server}
\title{Evidence Map Server Module}
\usage{
evidence_map_Server(
  id,
  citations_metadata = citations_for_dl,
  pico_data,
  x_axis_table,
  y_axis_table,
  legend_table,
  x_axis_specific_column = "",
  y_axis_specific_column = "",
  legend_specific_column = "",
  x_axis_group_column = "",
  y_axis_group_column = "",
  legend_group_column = "",
  x_dropdown_group_label = "",
  x_dropdown_specific_label = "",
  y_dropdown_group_label = "",
  y_dropdown_specific_label = "",
  legend_dropdown_group = "",
  legend_dropdown_specific = ""
)
}
\arguments{
\item{id}{The module identifier. Used to namespace the module's UI and server components.}

\item{citations_metadata}{A data table containing metadata for citations.65}

\item{x_axis_table}{A data frame containing data related to the x-axis of the evidence map.}

\item{y_axis_table}{A data frame containing data related to the y-axis of the evidence map.}

\item{legend_table}{A data frame containing data for the legend categories.}

\item{x_axis_specific_column}{A string specifying the column name in \code{x_axis_table}
for filtering specific x-axis values.}

\item{y_axis_specific_column}{A string specifying the column name in \code{y_axis_table}
for filtering specific y-axis values.}

\item{legend_specific_column}{A string specifying the column name in \code{legend_table}
for filtering specific legend values.}

\item{x_axis_group_column}{A string specifying the column name in \code{x_axis_table}
for filtering grouped x-axis values.}

\item{y_axis_group_column}{A string specifying the column name in \code{y_axis_table}
for filtering grouped y-axis values.}

\item{legend_group_column}{A string specifying the column name in \code{legend_table}
for filtering grouped legend values.}

\item{x_dropdown_group_label}{A string for the label of the grouped x-axis dropdown filter.}

\item{x_dropdown_specific_label}{A string for the label of the specific x-axis dropdown filter.}

\item{y_dropdown_group_label}{A string for the label of the grouped y-axis dropdown filter.}

\item{y_dropdown_specific_label}{A string for the label of the specific y-axis dropdown filter.}

\item{legend_dropdown_group}{A string for the label of the grouped legend dropdown filter.}

\item{legend_dropdown_specific}{A string for the label of the specific legend dropdown filter.}
}
\value{
None. This is a server-side Shiny module and does not return a value. It manages
reactive inputs, updates to filters, and rendering of the evidence map plot within a Shiny app.
}
\description{
This Shiny module server function handles the rendering and interactions for an evidence map,
including dropdown filters and the plot itself. It allows users to dynamically filter data
and generate a plotly evidence map.
}
\details{
This module supports an evidence map UI by dynamically creating and responding to user inputs
for filtering data by x-axis, y-axis, and legend categories. The module handles rendering
the corresponding plotly evidence map based on the filtered data. It also includes functionality
for resetting filters and managing state across sessions.
}
