% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{evidence_map_UI}
\alias{evidence_map_UI}
\title{Evidence Map UI Module}
\usage{
evidence_map_UI(id, switch_label = "")
}
\arguments{
\item{id}{The module identifier. Used to namespace the module's UI components.}

\item{switch_label}{A string for the label of the comparison mode toggle switch.}
}
\value{
A \code{tagList} containing the UI components for the evidence map module.
}
\description{
This Shiny module UI function creates the user interface for the evidence map.
It includes dropdown filters, a material switch for comparison mode, action buttons
for rendering and resetting the plot, a plotly evidence map, and a data table of selected studies.
}
\details{
The evidence map UI module provides dropdown filters, a comparison toggle, and action buttons for rendering and resetting the plot.
It includes a plotly evidence map with error handling and a data table of selected studies,
all organized into three boxes: Filters and Controls, Evidence Map, and Selected Studies.
}
