% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{extract_term_syrf_annotation}
\alias{extract_term_syrf_annotation}
\title{Extract and Clean Terms from SyRF Annotations}
\usage{
extract_term_syrf_annotation(
  annotation_df = NULL,
  annotated_column = NULL,
  res_filename = "term_list.csv",
  file_path = NULL
)
}
\arguments{
\item{annotation_df}{The syrf dataframe with annotation.}

\item{annotated_column}{Column with the annotated content.}

\item{res_filename}{Filename of result.}

\item{file_path}{Folder where result should be saved.}
}
\value{
This returns  a list of cleaned terms and also saves it in a csv file
}
\description{
This function processes a SyRF annotation dataframe to extract, clean, and
standardize a list of terms from a specified annotation column. It is designed
to handle semi-structured text where terms might be separated by various
delimiters or grouped within brackets. The function performs several cleaning
operations, removes stop words, and filters out very short terms to produce a
clean term list. The final list is saved to a CSV file and returned as a
character vector.
}
\examples{
\dontrun{
# Example usage:
extract_term_syrf_annotation(
annotation_df = syrf_df,
annotated_column = "Collected.sample._48b9.a2b3.32eeeb0b43de_Answer",
res_filename = "clean_term_list.csv")
}

}
