% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{field_codes_scopus}
\alias{field_codes_scopus}
\title{Field conversion codes for scopus imports}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 55 rows and 4 columns.
}
\usage{
data(field_codes_scopus)
}
\description{
Field conversion codes for scopus imports
}
\author{
Sean Smith \email{ssmith49@ed.ac.uk}
}
\keyword{datasets}
