% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{format_cols}
\alias{format_cols}
\title{Format Columns for SOLES Search Data}
\usage{
format_cols(df)
}
\arguments{
\item{df}{A dataframe to be formatted for SOLES search data.}
}
\value{
A formatted dataframe containing the required SOLES columns with standardized case.
}
\description{
This function processes a dataframe to retain relevant columns required for SOLES and standardizes the case of
specific columns to ensure consistent formatting.
}
