% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{format_tiab}
\alias{format_tiab}
\title{Format title and abstract character strings}
\usage{
format_tiab(df)
}
\arguments{
\item{df}{The SOLES dataframe to be formatted (e.g., \code{unique_citations} or \code{retrieved_citations})}
}
\value{
This function returns the formatted dataframe
}
\description{
This function performs various formatting operations on the title and abstract columns of a SOLES dataframe. These include: removal of leading and trailing whitespace, removal of double backslashes,
replacement of 2+ spaces with single space and removal of leading characters (en dash, em dash, hyphen, colon, dot).
}
\examples{
\dontrun{
# Example usage:
df \%>\% format_tiab()
}

}
