% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ft.R
\name{ft_epmc}
\alias{ft_epmc}
\title{Retrieve XML documents from EuropePMC using DOI}
\usage{
ft_epmc(doi, uid, ft_path, ft_name_style = "doi")
}
\arguments{
\item{doi}{character string containing DOI}

\item{uid}{optional unique ID for file naming}

\item{ft_path}{folder where XMLs should be downloaded}

\item{ft_name_style}{naming style for XMLs; options = c("doi", "uid")}
}
\value{
dataframe summary of retrieval
}
\description{
Retrieve XML documents from EuropePMC using DOI
}
