% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ft.R
\name{ft_unpaywall}
\alias{ft_unpaywall}
\title{Retrieve PDF documents from Unpaywall using DOI}
\usage{
ft_unpaywall(doi, uid, unpaywall_email, ft_path, ft_name_style = "doi")
}
\arguments{
\item{doi}{character string containing DOI}

\item{uid}{optional unique ID for file naming}

\item{unpaywall_email}{your email, required for retrieve from Unpaywall}

\item{ft_path}{folder where PDFs should be downloaded}

\item{ft_name_style}{naming style for PDFs; options = c("doi", "uid")}
}
\value{
dataframe summary of retrieval
}
\description{
Retrieve PDF documents from Unpaywall using DOI
}
