% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ft.R
\name{ft_wiley}
\alias{ft_wiley}
\title{Retrieve PDF documents from Wiley using DOI}
\usage{
ft_wiley(doi, uid, wiley_token, ft_path, ft_name_style = "doi")
}
\arguments{
\item{doi}{character string containing DOI}

\item{uid}{optional unique ID for file naming}

\item{wiley_token}{API key required for text data mining on Wiley texts}

\item{ft_path}{folder where PDFs should be downloaded}

\item{ft_name_style}{naming style for PDFs; options = c("doi", "uid")}
}
\value{
dataframe summary of retrieval
}
\description{
Retrieve PDF documents from Wiley using DOI
}
