% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crossref_retractions.R
\name{get_crossref_retractions}
\alias{get_crossref_retractions}
\title{Retrieve retraction information from Crossref}
\usage{
get_crossref_retractions(con, retraction_db = NULL)
}
\arguments{
\item{con}{connection to db}

\item{retraction_db}{dataframe containing CrossRef retraction database; will retrieve if NULL}
}
\value{
Updated "retraction_tag" table
}
\description{
This function checks DOIs against the Crossref retraction database.
}
\examples{
\dontrun{
# Example usage:
get_crossref_retractions(con)
}
}
