% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ft.R
\name{get_ft}
\alias{get_ft}
\title{Retrieve full text documents}
\usage{
get_ft(
  con,
  pdf_source = c("unpaywall", "crossref", "wiley"),
  xml_source = c("epmc", "elsevier"),
  ft_path,
  retMax = 500,
  check_failed = FALSE,
  timespan = NULL,
  wiley_token = NULL,
  elsevier_token = NULL,
  unpaywall_email = NULL
)
}
\arguments{
\item{con}{Database connection}

\item{pdf_source}{where to retrieve PDFs from, options = c("unpaywall", "crossref", "wiley") OR NULL}

\item{xml_source}{where to retrieve XMLs from, options = c("epmc", "elsevier") OR NULL}

\item{ft_path}{folder where full texts should be downloaded to}

\item{retMax}{maximum number of full texts to attempt to retrieve; max = 500; dault = 500}

\item{check_failed}{logical, whether to check DOIs that previously failed; default = FALSE}

\item{timespan}{timespan for check_failed in format digit(week/month) e.g. "2month"}

\item{wiley_token}{API token required to retrieve from Wiley}

\item{elsevier_token}{API token required to retrieve from Elsevier}

\item{unpaywall_email}{Your email, required to retrieve from Unpaywall}
}
\value{
summary of retrieval wrritten to database
}
\description{
Retrieve full texts as PDF or XML for SOLES.
}
