% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_missing_metadata.R
\name{get_missing_abstracts}
\alias{get_missing_abstracts}
\title{Fill in missing abstracts via CrossRef}
\usage{
get_missing_abstracts(
  citations,
  abstract_col = "abstract",
  doi_col = "doi",
  id_col = "uid"
)
}
\arguments{
\item{citations}{a dataframe containing records missing abstract text data}

\item{abstract_col}{the name of the column that should contain the abstract text. Default: \code{abstract}}

\item{doi_col}{the name of the column that contains the DOI. Default: \code{doi}}

\item{id_col}{a unique identifier column. For SyRF output this could be \code{CustomId}. Default: \code{uid}}
}
\value{
The same dataframe entered as citations with additional abstracts entered where these could be retrieved by CrossRef
}
\description{
Search records in CrossRef using DOI to attempt to retrieve missing abstract text.
}
