% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_missing_metadata.R
\name{get_missing_dois}
\alias{get_missing_dois}
\title{Fill in missing DOIs}
\usage{
get_missing_dois(citations)
}
\arguments{
\item{citations}{citations you want to find dois for}
}
\value{
Dataframes with additional DOIs found by OpenAlex
}
\description{
Uses title matching to retrieve DOI information from OpenAlex
}
