% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup.R
\name{get_new_unique}
\alias{get_new_unique}
\title{Get new unique citations}
\usage{
get_new_unique(con, new_citations, prev_months = 2)
}
\arguments{
\item{con}{connection to db}

\item{new_citations}{new citations identified}

\item{prev_months}{how many previous months of citations do you want to dedup against?}
}
\value{
recent citations
}
\description{
Get new unique citations
}
