% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_openalex_update.R
\name{get_openalex_update}
\alias{get_openalex_update}
\title{Update citation count and retraction data from OpenAlex.}
\usage{
get_openalex_update(con)
}
\arguments{
\item{con}{connection to db}
}
\value{
Tables "citation_count_tag", "retraction_tag" will be overwritten with updated with data. .
}
\description{
This function retrieves and updates meta-data on DOIs where metadata was last retrieved over 6 months ago.
}
\examples{
\dontrun{
# Example usage:
get_openalex_update(con)
}
}
