% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ror_coords.R
\name{get_ror_coords}
\alias{get_ror_coords}
\title{Retrieve longitude and latitude coordinates from ROR for institutions}
\usage{
get_ror_coords(con)
}
\arguments{
\item{con}{connection to db}
}
\value{
Tables "ror_coords" will be updated with n number of new rows containing metadata. If no data is retrieved for these n citations, "Unknown" will be returned in the relevant columns.
}
\description{
This function retrieves longitude and latitude coordinates.
}
\examples{
\dontrun{
# Example usage:
get_ror_coords(con)
}
}
