% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_studies.R
\name{get_studies_to_screen}
\alias{get_studies_to_screen}
\title{Get unscreened studies from database'}
\usage{
get_studies_to_screen(con, classify_NA = FALSE, project_name, classifier_name)
}
\arguments{
\item{con}{connection to db}

\item{classify_NA}{boolean true/false based on whether you want to retain missing abstracts - the machine
does not perform well with missing abstracts}

\item{project_name}{name of soles project}

\item{classifier_name}{name of classifier e.g. in vivo}
}
\value{
citations which have not been screened
}
\description{
Get unscreened studies from database'
}
