% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_syrf_sample}
\alias{get_syrf_sample}
\title{Get a sample of citations for screening in SyRF}
\usage{
get_syrf_sample(df, sample_size = 2000, abstracts_only = TRUE)
}
\arguments{
\item{df}{A dataframe containing citation data.}

\item{sample_size}{The size of the random sample to be obtained.}

\item{abstracts_only}{Logical, indicating whether to filter only rows with abstracts.}
}
\value{
A dataframe containing the random sample with specified columns.
}
\description{
This function takes a dataframe and returns a random sample of specified size.
Optionally, it can filter the dataframe to include only rows with abstracts.
This function also writes the resulting dataframe to a csv file and saves in working directory.
}
\examples{
\dontrun{
# Get a random sample without removing rows containing no abstract
result <- get_syrf_sample(my_data, sample_size = 1000, abstracts_only = FALSE)
}

}
