% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_manual.R
\name{manual_upload}
\alias{manual_upload}
\title{Import Search Results from Manual Searches of Bibliographic Databases}
\usage{
manual_upload(paths, source)
}
\arguments{
\item{paths}{a character vector of file paths to the manual search result files}

\item{source}{the source database for the search results. Choices: "scopus", "embase", "psychinfo", "medline", "wos", "pubmed", or "endnote")}
}
\value{
a dataframe containing the search results from the specified source, formatted for SOLES
}
\description{
Read search results manually downloaded from bibliographic databases into R for use in SOLES workflows.
Columns are formatted for SOLES integration.
}
\details{
Data can be imported from Ovid databases (Scopus, Embase, PsychINFO, or MEDLINE) in \code{.bib}, \code{.ris}, and \code{.txt} format;
Web of Science in \code{.bib} or \code{.ris} format; and PubMed in \code{.bib} format.
Additionally, data from EndNote can be imported via EndNote XML file. No other XML file formats are permitted.
}
