% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{medline2df}
\alias{medline2df}
\title{Convert a RISmed::Medline object into a SOLES pubmed search data frame}
\usage{
medline2df(M)
}
\arguments{
\item{M}{an object of class MedLine (package "RISmed")}
}
\value{
A dataframe formatted for the SOLES workflow
}
\description{
Convert a RISmed::Medline object into a SOLES pubmed search data frame
}
