% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rncbo_dictionary.R
\name{NCBO_get_fields_from_JSON}
\alias{NCBO_get_fields_from_JSON}
\title{NCBO_get_fields_from_JSON}
\usage{
NCBO_get_fields_from_JSON(CampaignJSON)
}
\arguments{
\item{CampaignJSON}{A JSON object returned from an NCBO REST Search query.}
}
\value{
A character vector containing the ID, notation, definition, and synonyms for a term. If a field is not found, it returns NA. If there's an error, the ID will be "ERROR!".
}
\description{
This function parses the JSON output from an NCBO REST Search query and extracts specific fields: ID, notation, definition, and synonyms.
}
