% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rncbo_dictionary.R
\name{NCBO_get_parents_from_JSON}
\alias{NCBO_get_parents_from_JSON}
\title{NCBO_get_parents_from_JSON}
\usage{
NCBO_get_parents_from_JSON(CampaignJSON, apikey)
}
\arguments{
\item{CampaignJSON}{The JSON object from the NCBO Annotator.}

\item{apikey}{Your NCBO API key.}
}
\value{
A character vector of parent terms' preferred labels, concatenated with "|". If no parents are found, it returns NA.
}
\description{
This function takes a JSON object from the NCBO Annotator and an API key, extracts the URL for the parent terms, fetches them, and returns the preferred labels of the parent terms.
}
