% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rncbo_dictionary.R
\name{NCBO_mapper}
\alias{NCBO_mapper}
\title{NCBO_mapper}
\usage{
NCBO_mapper(term_file, ontology_file, result_file, apikey)
}
\arguments{
\item{term_file}{The path to a CSV file with the terms to map (one term per line).}

\item{ontology_file}{The path to a CSV file with the ontology acronyms to use for mapping (one acronym per line).}

\item{result_file}{The path to the output CSV file where the mappings will be saved.}

\item{apikey}{Your NCBO API key.}
}
\value{
This function does not return a value but writes the mapping results to \code{result_file} and a list of non-mapped terms to a file with the suffix "NonMapped.csv".
}
\description{
This function reads a list of terms and a list of ontologies, and for each term, it uses the NCBO Recommender, Search, and Annotator APIs to find mappings in the specified ontologies. The results are written to a CSV file. This version is refactored to be more efficient by processing all terms and writing to disk only once.
}
