% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rncbo_dictionary.R
\name{NCBO_ontologies}
\alias{NCBO_ontologies}
\title{NCBO_ontologies}
\usage{
NCBO_ontologies(
  terms_list,
  apikey,
  file_path = "",
  n_sample = 5,
  n_terms = 100,
  topn_onto = 5
)
}
\arguments{
\item{terms_list}{A character string with the path to a file containing a list of terms, one term per line.}

\item{apikey}{Your NCBO API key.}

\item{file_path}{The path to the directory where the output file with the list of ontologies will be saved.}

\item{n_sample}{The number of samples to create from the terms list. Default is 5.}

\item{n_terms}{The number of terms to include in each sample. Default is 100.}

\item{topn_onto}{The number of top ontologies to select from the recommendations for each sample. Default is 5.}
}
\value{
A character vector with the unique acronyms of the recommended ontologies. It also writes a file named "ontology_list.csv" in the specified \code{file_path} with the list of recommended ontologies.
}
\description{
This function takes a list of terms and uses the NCBO Recommender API to suggest relevant ontologies. It samples the terms, sends them to the API, and collects the top recommended ontologies.
}
\examples{
\dontrun{
NCBO_ontologies(
 terms_list = TERM_LIST,
 apikey = YOUR_API_KEY,
 file_path = "data/dictionary/NCBO"
)
}
}
