% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rncbo_dictionary.R
\name{NCBO_recommender_from_JSON}
\alias{NCBO_recommender_from_JSON}
\title{NCBO_recommender_from_JSON}
\usage{
NCBO_recommender_from_JSON(term, CampaignJSON)
}
\arguments{
\item{term}{The term that was sent to the recommender.}

\item{CampaignJSON}{The JSON object returned by the NCBO Recommender API.}
}
\value{
A data frame with the recommended ontologies for the term. The data frame has the following columns: "Term", "Found_term", "Score", "Ontology", "Ontology_URI". Returns NULL if there are errors or no results.
}
\description{
This function processes a JSON object from the NCBO Recommender API and extracts the recommended ontologies, their scores, and other details for a given term.
}
