% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_oa_tag.R, R/oa_tag.R
\name{oa_tag}
\alias{oa_tag}
\title{Extract open access status from Unpaywall}
\usage{
oa_tag(con, email, n = 100)

oa_tag(con, email, n = 100)
}
\arguments{
\item{con}{connection to db}

\item{email}{email address to use for Unpaywall}

\item{n}{number of citations to tag for open access}
}
\description{
This function retrieves meta-data.

This function retrieves meta-data.
}
