% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods_tag.R
\name{ods_tag}
\alias{ods_tag}
\title{Extract open data and code statements using ODDPub}
\usage{
ods_tag(con, path)
}
\arguments{
\item{con}{connection to db}

\item{path}{folder containing full texts}
}
\description{
This function evaluates reporting of open data and code statements and extracts those statements.
}
