% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{pico_dropdown_Server}
\alias{pico_dropdown_Server}
\title{PICO Dropdown Module Server}
\usage{
pico_dropdown_Server(id, table, column1, column2, column3, column4, filter_no)
}
\arguments{
\item{id}{The module identifier.}

\item{table}{The data table containing the columns used for filtering.}

\item{column1}{The name of the first column to be used for generating the first dropdown.}

\item{column2}{The name of the second column to be used for generating the second dropdown.}

\item{column3}{The name of the third column to be used for generating the third dropdown.}

\item{column4}{The name of the fourth column to be used for generating the fourth dropdown.}

\item{filter_no}{The number of dropdown menus required.}
}
\description{
This Shiny module server function handles the dynamic behavior of reactive dropdown menus based on PICO elements.
}
