% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{pico_dropdown_UI}
\alias{pico_dropdown_UI}
\title{PICO Dropdown Module UI}
\usage{
pico_dropdown_UI(
  id,
  column1,
  column2,
  column3,
  column4,
  label1,
  label2,
  label3,
  label4,
  filter_no
)
}
\arguments{
\item{id}{The module identifier.}

\item{column1}{The name of the first column to use for generating the dropdown.}

\item{column2}{The name of the second column to use for generating the second dropdown.}

\item{column3}{The name of the third column to use for generating the third dropdown.}

\item{column4}{The name of the fourth column to use for generating the fourth dropdown.}

\item{label1}{The label shown above the first dropdown.}

\item{label2}{The label shown above the second dropdown.}

\item{label3}{The label shown above the third dropdown.}

\item{label4}{The label shown above the fourth dropdown.}

\item{filter_no}{The number of dropdown menus required.}
}
\description{
This Shiny module creates reactive dropdown menus for filtering based on PICO elements.
}
