% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{pico_multi_select_Server}
\alias{pico_multi_select_Server}
\title{Multi-Select Bar Chart for PICO Tabs}
\usage{
pico_multi_select_Server(id, multi_select = TRUE, table, column, column2, text)
}
\arguments{
\item{id}{The module identifier.}

\item{multi_select}{Specifies whether to use 1 or 2 dropdown menus.}

\item{table}{The data table for generating the chart.}

\item{column}{The name of the first column for generating the chart.}

\item{column2}{The name of the second column (if multi-select is TRUE and 2 dropdown menus are required).}

\item{text}{The text to display below the chart stating the origin of the data.}
}
\description{
This is the server function for the PICO multi-select module. It generates a bar chart by year based on the inputs.
}
