% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pico_tag.R
\name{pico_tag}
\alias{pico_tag}
\title{Tag citations using regex}
\usage{
pico_tag(
  con,
  tag_type,
  tag_method,
  tag_main_category = "all",
  ignore_case = FALSE,
  extract_strings = TRUE
)
}
\arguments{
\item{con}{connection to db - note that this must contain tables of the format 'type_tag' and dictionaries of the type type_dictionary}

\item{tag_type}{tag type related to db table headers e.g. outcome, model, or intervention}

\item{tag_method}{full_text OR tiabkw}

\item{tag_main_category}{tag main category from pico ontology table, leave as default all if you want all for that type}

\item{ignore_case}{do you want to ignore case for regexes?}

\item{extract_strings}{do you want to extract regex matches from text?}
}
\description{
This function tags included citations with different regex dictionaries, using the full text and title/abstract/keyword fields
separately. Finally, tags are added to the relevant database tables
}
