% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{pie_completion_Server}
\alias{pie_completion_Server}
\title{Pie Chart Showing Tagged Completion Percentage}
\usage{
pie_completion_Server(
  id,
  table,
  identifier,
  included_studies,
  colour_not_complete,
  colour_complete,
  incl_missing_doi = FALSE
)
}
\arguments{
\item{id}{The module identifier.}

\item{table}{The data table to use for generating the chart.}

\item{identifier}{The name of the column to use for generating the chart.}

\item{included_studies}{The table used for bringing in included studies.}

\item{colour_not_complete}{The colour for the section of pie showing "not complete".}

\item{colour_complete}{The colour for the section of pie showing "complete".}

\item{incl_missing_doi}{For tables which are linked by DOI and also want to show \% without a DOI in the chart}
}
\description{
This shiny module generates a box containing a pie chart showing the tagged completion percentage of a given dataframe. This can also
include the number of studies missing a doi.
}
