% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{pie_completion_UI}
\alias{pie_completion_UI}
\title{Box Containing Pie Chart Indicating Completion}
\usage{
pie_completion_UI(id, title, theme, spinner_colour, info_text)
}
\arguments{
\item{id}{The module identifier.}

\item{title}{The title of the box/pie chart.}

\item{theme}{The status or colour theme of the box.}

\item{spinner_colour}{The colour of the loading spinner.}

\item{info_text}{The text to be shown in the box above the pie chart.}
}
\description{
This shiny module generates a box containing a pie chart showing the tagged completion percentage of a given dataframe.
}
