% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{plot_interpret_Server}
\alias{plot_interpret_Server}
\title{Info Box Module Server}
\usage{
plot_interpret_Server(id)
}
\arguments{
\item{id}{The module identifier.}
}
\description{
This is the server function for the info box module. It currently doesn't have any specific functionality.
}
