% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{plot_interpret_UI}
\alias{plot_interpret_UI}
\title{Info Box Module UI}
\usage{
plot_interpret_UI(
  id,
  title = "How to interpret this plot",
  info_text = "",
  theme
)
}
\arguments{
\item{id}{The module identifier.}

\item{title}{The title of the info box.}

\item{info_text}{The text to be shown in the info box.}

\item{theme}{The color theme of the info box.}
}
\description{
This Shiny module creates an info box with a title and information text.
}
