% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{process_embase}
\alias{process_embase}
\title{Process Embase references}
\usage{
process_embase(path)
}
\arguments{
\item{path}{The file path of the data.}
}
\value{
A processed data frame.
}
\description{
This function processes references from Embase.
}
\details{
This function reads references from the specified file using the "ovid" tag naming convention.
It performs various transformations on the data, such as uniting title and booktitle columns,
formatting pages, creating a unique identifier (uid) based on the article number (AN), and
setting the data source to "embase".
}
\examples{
\dontrun{
# Example usage:
embase_data <- process_embase("path/to/embase_data.csv")
}

}
