% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_manual.R
\name{process_endnote}
\alias{process_endnote}
\title{Process Manually Uploaded Search results from Endnote XML file}
\usage{
process_endnote(path)
}
\arguments{
\item{path}{the file path to the search results}
}
\value{
a processed dataframe containing search results
}
\description{
Internal function used to process EndNote exports as XML an format for SOLES.
}
