% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{process_medline}
\alias{process_medline}
\title{Process Medline references}
\usage{
process_medline(path)
}
\arguments{
\item{path}{The file path of the data.}
}
\value{
A processed data frame.
}
\description{
This function processes references from Medline.
}
\details{
This function reads references from the specified file using the "ovid" tag naming convention.
It performs various transformations on the data, such as uniting title and booktitle columns,
formatting pages, creating a unique identifier (uid) based on the article_id, and setting
the data source to "medline".
}
\examples{
\dontrun{
# Example usage:
medline_data <- process_medline("path/to/medline_data.csv")
}

}
