% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_manual.R
\name{process_wos}
\alias{process_wos}
\title{Process Manually Uploaded Search results from Web of Science}
\usage{
process_wos(path)
}
\arguments{
\item{path}{the file path to the search results}
}
\value{
a processed dataframe containing search results
}
\description{
Internal function used to process Web of Science search results an format for SOLES.
}
\details{
This function reads references from the specified file in BibTeX or RIS format using the "wos" database source.
The function checks the file extension before processing accordingly.
It adjusts column names according to the field codes, removes blank columns, adds source information,
creates a unique identifier (uid) by modifying the existing uid column, and sets the data source to "wos".
}
\examples{
\dontrun{
# Example usage:
wos_data <- process_wos("path/to/wos_data.bib")
wos_data <- process_wos("path/to/wos_data.ris")
}

}
