% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{process_xml}
\alias{process_xml}
\title{Process XML File}
\usage{
process_xml(path)
}
\arguments{
\item{path}{The file path to the XML file.}
}
\value{
A data frame containing parsed bibliographic information.
}
\description{
This function parses an XML file containing bibliographic information and extracts relevant fields.
}
