% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{pubmed_search}
\alias{pubmed_search}
\title{Get pubmed results}
\usage{
pubmed_search(query, timespan, retMax = 5000)
}
\arguments{
\item{query}{search query}

\item{timespan}{time frame for search options e.g. 1week, 2week, 1month}

\item{retMax}{maximum number of citations to retrieve, maximum is 9999}
}
\value{
pubmed search results
}
\description{
This function retrieves search results from pubmed
}
