% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{read_search}
\alias{read_search}
\title{Load a search export file, determine its source and extract date from filename.}
\usage{
read_search(file)
}
\arguments{
\item{file}{A file containing search results. Valid sources are all literature databases supported in soles:
pubmed, wos, scopus, medline, ovid, embase and psychinfo.}
}
\description{
Load a search export file, determine its source and extract date from filename.
}
\keyword{internal}
