% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_tag.R
\name{regex_tag}
\alias{regex_tag}
\title{Tag records using regex}
\usage{
regex_tag(
  con,
  tag_type,
  tag_method,
  tag_main_category = "all",
  retMax = 1000,
  ignore_case = FALSE,
  extract_strings = TRUE
)
}
\arguments{
\item{con}{valid database connection}

\item{tag_type}{type of tagging, e.g. species or model}

\item{tag_method}{method of tagging, either tiab or fulltext}

\item{tag_main_category}{set to all to tag all categories or sepcific category}

\item{retMax}{maximum numer of recrods to tag; default 1000; max 5000}

\item{ignore_case}{logical; whether to globally ignore case in regex}

\item{extract_strings}{logical; whether to extract strings where regex matches occur}
}
\value{
database appended with newly tagged records
}
\description{
Tag SOLES records using title and abstract or full (PDF or XML) text
}
