% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{rm_html_tags}
\alias{rm_html_tags}
\title{Remove HTML tags from text}
\usage{
rm_html_tags(string)
}
\arguments{
\item{string}{The input text for formatting}
}
\value{
This function returns the formatted input string, with HTML tags removed.
}
\description{
A function that can be used to remove HTML tags from titles and abstracts
}
\details{
Currently, only basic text formatting tags are identified and removed. See: https://www.w3schools.com/html/html_formatting.asp
}
\examples{
\dontrun{
# Example usage:
df \%>\% mutate(across(c(title, abstract), ~ rm_html_tags(.)))
}

}
