% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob.R
\name{rob_tag}
\alias{rob_tag}
\title{Tag studies by risk of bias}
\usage{
rob_tag(con, max_file_size = 5e+05, num_cores = 10, py_path = "")
}
\arguments{
\item{con}{connection to db}

\item{max_file_size}{maximum file size in bites}

\item{num_cores}{number of cores to use for parallelisation}

\item{py_path}{path to python risk of bias tool e.g. "/opt/sharedFolder/rob/rob-app/"}
}
\description{
Tag studies by risk of bias
}
