% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_studies.R
\name{run_k_fold}
\alias{run_k_fold}
\title{Run K-Fold Cross-Validation for Machine Learning}
\usage{
run_k_fold(
  con,
  project_name = as.character(),
  sample = FALSE,
  sample_number = as.numeric(),
  file_id = "",
  fold_number = 5,
  repeats = 3,
  training_prop = 0.8,
  classifier_name = NULL
)
}
\arguments{
\item{con}{A database connection object.}

\item{project_name}{A string specifying the project name. Used for naming output files and directories.}

\item{sample}{Boolean true/false based on whether you want to use all screening decisions or take a random sample.}

\item{sample_number}{An integer specifying the number of studies you want in the sample.}

\item{file_id}{A string to be used as a file id to distinguish results, for repeat runs of this function on the same day.}

\item{fold_number}{An integer specifying the number of folds for cross-validation. Default is 5.}

\item{repeats}{An integer specifying how many repeats of the k-fold validation}

\item{training_prop}{A decimal between 0 and 1. The proportion of all of the screening decisions to be used for training. Set to 1 to use all studies for k-fold with no validation. Default is 0.8.}

\item{classifier_name}{A string specifying the classifier name to be used, taken from the study_classification table (e.g. "in-vivo", "in-vitro", "clinical")}
}
\value{
The function does not return an R object. It writes the results to the \code{k-fold-validation} directory.
}
\description{
This function performs K-fold cross-validation. It retrieves data
from a database, splits it into \code{k} folds, trains and tests a classifier for each fold, and
computes performance metrics. Results are saved to CSV files for further analysis.
}
\details{
This function:
\enumerate{
\item Retrieves included and excluded screening decisions from the database.
\item Shuffles the data and splits it into the specified number of folds.
\item For each iteration:
\itemize{
\item Assigns a test set while others are used for training.
\item Runs the specified machine learning classifier via an external API.
\item Computes and saves performance metrics and test scores.
}
}

Results are saved in a directory named \code{k-fold-validation} in the current working directory.
}
\examples{
\dontrun{
# Example usage:
run_k_fold(
  con = db_con,
  project_name = "MS-SOLES",
  sample = FALSE,
  file_id = "_1",
  fold_number = 5,
  repeats = 3,
  classifier_name = "animal"
)
}
}
