% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_studies.R
\name{run_ml}
\alias{run_ml}
\title{Run machine learning to get included studies and send them to soles database}
\usage{
run_ml(con, training_set, unscreened_set, project_name, classifier_name)
}
\arguments{
\item{con}{connection to db}

\item{training_set}{set of screened citations with 1/0 LABEL}

\item{unscreened_set}{set of unscreened citations LABEL 99}

\item{project_name}{name of soles project}

\item{classifier_name}{name of classifier that describes what it does e.g. "screening" or "in vivo"}
}
\value{
citations which have not been screened
}
\description{
Run machine learning to get included studies and send them to soles database
}
