% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_studies.R
\name{run_ml_kfold_val}
\alias{run_ml_kfold_val}
\title{Run ML using the training and validation sets from K-fold}
\usage{
run_ml_kfold_val(con, training_set, validation_set, project_name, file_id = "")
}
\arguments{
\item{con}{connection to db}

\item{training_set}{set of screened citations with 1/0 LABEL}

\item{validation_set}{set of citations to be screened for validation LABEL 99}

\item{project_name}{name or project identifier}

\item{file_id}{A string to be used as a file id to distinguish results, for repeat runs of this function on the same day.}
}
\value{
A CSV results file of probability scores
}
\description{
This function takes the stratified training/validation splits from the k-fold function and re-trains the algorithm using
all of the training data to get scores for the final validation set.
}
