% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_api.R
\name{scopus_search}
\alias{scopus_search}
\title{Search Scopus and retrieve bibliographic data using scopusAPI}
\usage{
scopus_search(
  query = NULL,
  api_key = NULL,
  timespan = NULL,
  retMax = 2000,
  format_soles = TRUE
)
}
\arguments{
\item{query}{a character string containing a correctly syntaxed Scopus search}

\item{api_key}{a working Scopus API key}

\item{timespan}{a formatted character string defining the timespan you want to search}

\item{retMax}{The maximum number of records to retrieve, default is 2000, maximum is 5000}

\item{format_soles}{boolean, if set to TRUE will format search results for the SOLES workflow, default is TRUE}
}
\value{
a dataframe containing Scopus search results
}
\description{
A wrapper function for scopusAPI. Search Scopus using a query and retrieve results programmatically.
Requires \code{devtools::install_github(kaitlynhair/scopusAPI)}.
Requires an API key from Scopus (http://dev.elsevier.com/).
Use the function \code{usethis::edit_r_environ()} to add the key to your ~/.Renviron file.
The timespan should be formatted as a number followed by the word "week" or "month", e.g. "1month" or "2week".
}
\examples{
\dontrun{
# api_key should be stored in your r environment and not shared
query <- "TITLE-ABS-KEY(dementia OR \"memory loss\")"
scopus_result <- scopus_search(query, api_key = api_key, retMax = 500, format_soles = FALSE)
}
}
