% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{search_UI}
\alias{search_UI}
\title{Search Database Module UI}
\usage{
search_UI(id, table)
}
\arguments{
\item{id}{The module identifier.}

\item{table}{A dataframe containing the data for searching.}
}
\description{
This Shiny module creates the search page for the app.
}
