% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_soles}
\alias{setup_soles}
\title{This function reads in search files in the soles project directory and populates the \code{unique_citations} and \verb{retrieved_citations tables}.
\code{setup_soles()} expects that the required tables in the project database already exist. See \code{create_tables()}.}
\usage{
setup_soles(con, folder, master_node)
}
\arguments{
\item{con}{Connection to db}

\item{folder}{The name of the folder in the project directory where all search export files are saved.}

\item{master_node}{OSF master node}
}
\description{
This function reads in search files in the soles project directory and populates the \code{unique_citations} and \verb{retrieved_citations tables}.
\code{setup_soles()} expects that the required tables in the project database already exist. See \code{create_tables()}.
}
\details{
The function expects that within the main search folder, each sub-folder is named after the corresponding database source.
Valid sub-folder names are: pubmed, medline, wos, scopus, ovid, psychinfo, embase.
The function also allows users to include a date, following the naming convention \code{source_date}, where \code{source}
is the literature database and \code{date} corresponds to the date the search was conducted in \code{ddmmyyyy} format.
If a date is included in the folder name, this will populate the \code{date} field of unique citations.

For example, the following would be valid folder names: \code{pubmed_010123}, \code{scopus_151022}.
}
\examples{
\dontrun{
# Example usage:
setup_soles(con, "path/to/search/folder")
setup_soles(con, "manual_search")
}

}
