% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{sunburstSever}
\alias{sunburstSever}
\title{Generate a Sunburst Plot Module}
\usage{
sunburstSever(id, data)
}
\arguments{
\item{id}{The module identifier.}

\item{data}{The dataset containing the information for visualization.}
}
\description{
This Shiny module generates a sunburst plot based on the provided dataset.
}
