% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{sunburstUI}
\alias{sunburstUI}
\title{Generate a Sunburst Plot UI}
\usage{
sunburstUI(id, title = "", theme = "", spinner_colour)
}
\arguments{
\item{id}{The module identifier.}

\item{title}{The title of the tab.}

\item{theme}{The color (status) of the tab.}

\item{spinner_colour}{The color of the loading spinner.}
}
\description{
This Shiny module creates a sunburst plot.
}
