% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{syrf_decisions_to_db}
\alias{syrf_decisions_to_db}
\title{Write the screening decision from SyRF to the database}
\usage{
syrf_decisions_to_db(
  con,
  file = "",
  with_annotations = FALSE,
  classifier_name = ""
)
}
\arguments{
\item{con}{A connection to the target database.}

\item{file}{The file path to the SyRF screening data in CSV format.}

\item{with_annotations}{Logical, indicating whether the SyRF file contains annotations (TRUE/FALSE).}

\item{classifier_name}{The name of the classifier to train the data on.}
}
\value{
The function does not return a value but writes the processed data to the database.
}
\description{
This function reads a CSV file containing SyRF screening decisions and uploads
the data to a specified database. If the data includes annotations, a message
is displayed, indicating that manual cleaning is required.
}
\examples{
\dontrun{
  syrf_decisions_to_db(con = your_database_connection,
                       file = "path/to/your/data.csv",
                       with_annotations = FALSE,
                       classifier_name = "my-project")
}

}
