% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{tag_update_oa_od_rob}
\alias{tag_update_oa_od_rob}
\title{Update Open Access, Open Data, and Risk of Bias tags}
\usage{
tag_update_oa_od_rob(con, rob_max_file_size = 5e+05, email = "")
}
\arguments{
\item{con}{A database connection object.}

\item{rob_max_file_size}{Maximum file size for RoB tagging.}

\item{email}{Email address for using Unpaywall.}
}
\value{
None. Database tables will be updated.
}
\description{
This function tags all of the citations available for open access, open data and risk of bias.
}
\details{
The function retrieves studies to tag for OA and OD, and tags them using respective services.
It also tags studies for RoB based on PDF files within a specified size range.
The tagging process continues until no new studies are tagged or an error occurs.
}
\examples{
\dontrun{
tag_update_oa_od_rob(rob_max_file_size = 500000, email = "example/email.com")
}
}
