% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ft.R
\name{wiley_ft}
\alias{wiley_ft}
\title{Wiley full text retrieval}
\usage{
wiley_ft(doi, uid, token, path)
}
\arguments{
\item{doi}{digital object identifier}

\item{uid}{unique identifier for paper}

\item{token}{API token}

\item{path}{where full texts should be stored}
}
\description{
This function retrieves full texts from Wiley journals
}
