% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{yearBarServer}
\alias{yearBarServer}
\title{Server Function for Percentage Yearly Bar Chart}
\usage{
yearBarServer(
  id,
  table,
  column,
  order = c("reported", "not reported", "unknown"),
  display = "reported",
  text = "",
  colours = c("#76A8C1", "#FFC076", "grey")
)
}
\arguments{
\item{id}{The module identifier.}

\item{table}{The data table used for chart generation.}

\item{column}{The name of the column to use for chart generation.}

\item{order}{The order of legend items. Default is c("reported", "not reported").}

\item{display}{The value of the column to display. Default is "reported".}

\item{text}{Additional annotation text for the chart.}

\item{colours}{Colors for the chart elements.}
}
\description{
This server function generates a percentage bar chart by year based on the input table, column name, and display option. The resulting chart displays the percentage of occurrences for each year and includes a legend with a color-coded representation of the selected column values.
}
