% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{yearBarServer_included_only}
\alias{yearBarServer_included_only}
\title{Server Function for Percentage Yearly Bar Chart}
\usage{
yearBarServer_included_only(id, table, column, text = "", colour = "#73D055FF")
}
\arguments{
\item{id}{The module identifier.}

\item{table}{The data table used for chart generation.}

\item{column}{The name of the column to use for chart generation.}

\item{text}{Additional annotation text for the chart.}

\item{colour}{Color for the chart elements.}
}
\description{
This server function generates a percentage bar chart by year based on the input table, column name, and display option. The resulting chart displays the percentage of occurrences for each year and includes a legend with a color-coded representation of the selected column values.
}
