% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{yearBarUI_included_only}
\alias{yearBarUI_included_only}
\title{Generate a Bar Chart by Year}
\usage{
yearBarUI_included_only(
  id,
  title = "",
  theme = "",
  spinner_colour = "#96c296",
  table
)
}
\arguments{
\item{id}{The module identifier.}

\item{title}{The title of the tab.}

\item{theme}{The color (status) of the tab.}

\item{spinner_colour}{The color of the spinner shown while the plot is generating.}

\item{table}{The input data table.}
}
\description{
This Shiny module creates a count bar chart by year based on a specified table and selected column. The resulting chart displays the count of occurrences of the values of the selected column for each year, along with a legend featuring a color-coded representation of those values. Intended for use within a Shiny application.
}
