% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pico_utils.R
\name{add_regex_soles}
\alias{add_regex_soles}
\title{This function takes in user-provided RegEx entries and cross-checks them against the main SOLES pico dictionary on OSF.
If the provided RegEx are syntactically valid and represent new concepts, they are optionally added to the OSF repository, as well as the SOLES project-specific pico tables. Note that all RegEx should be perl-compatible to be considered valid.}
\usage{
add_regex_soles(con, regexfile, master_node, add_node)
}
\arguments{
\item{con}{Connection to soles project db}

\item{regexfile}{The path to the csv file containing the RegEx. The file should contain the following columns: \code{name}, \code{regex}, \code{type}, \code{main_category}, \code{sub_category1}, and \code{sub_category2}. If the ontology columns are empty, NAs will be introduced.}

\item{master_node}{OSF node for project to retrieve.}

\item{add_node}{OSF node for project to retrieve.}
}
\description{
This function takes in user-provided RegEx entries and cross-checks them against the main SOLES pico dictionary on OSF.
If the provided RegEx are syntactically valid and represent new concepts, they are optionally added to the OSF repository, as well as the SOLES project-specific pico tables. Note that all RegEx should be perl-compatible to be considered valid.
}
\details{
To be able to add new entries to the main dictionary, users will be asked to provide metadata such as the domain of new regex entries and their initials. This information will be included in the log file on OSF.
}
\examples{
\dontrun{
# Example usage:
add_regex_soles(con, "path/to/new/regex.csv")
}

}
