% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_combine.R
\name{check_if_retrieved}
\alias{check_if_retrieved}
\title{Update retrieved citations table in SOLES AWS}
\usage{
check_if_retrieved(con, citations)
}
\arguments{
\item{con}{connection to an AWS database}

\item{citations}{a dataframe of all new citations retrieved}
}
\value{
dataframe of new citations, new citations appended to retrieved_citations AWS table
}
\description{
This function is part of the SOLES architexture and requires a valid SOLES AWS set up.
Check what records have been retrieved previously and add the newly retrieved citations to the \code{retrieved_citations} table in SOLES AWS.
}
