% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_combine.R
\name{combine_searches}
\alias{combine_searches}
\title{Combine SOLES-compatible search results into one dataframe}
\usage{
combine_searches(...)
}
\arguments{
\item{...}{one or more data frame objects to be combined}
}
\value{
dataframe of combined search results with SOLES-compatible column names
}
\description{
Combine results from manual and / or API searches in a SOLES-compatible format.
Each data frame should represent search results and can have different sets of columns.
The function automatically aligns these data frames by column names, appending missing columns as \code{NA}.
SOLES columns in output are: "uid", "source", "author", "year", "journal", "doi", "title",
"pages", "volume", "abstract", "isbn", "keywords",
"secondarytitle", "url", "date", "issn", "pmid", "ptype",
"source", "author_country", "number", "author_affiliation".
}
\examples{
\dontrun{
combined_result <- combine_searches(pubmed, wos, scopus)
}
}
