% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{create_regex}
\alias{create_regex}
\title{Create Regular Expressions}
\usage{
create_regex(file = "")
}
\arguments{
\item{file}{Path to the input Excel file.}
}
\value{
A new Excel file with additional columns containing regular expressions.
}
\description{
This function reads a file containing names and their corresponding alternate names,
then creates regular expressions based on these names and alternate names.
}
\examples{
\dontrun{
# Example usage:
create_regex("input_file.xlsx")
}

}
