% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{download_table_Server}
\alias{download_table_Server}
\title{Download table server}
\usage{
download_table_Server(id, table)
}
\arguments{
\item{id}{The module identifier.}

\item{table}{}
}
\description{
This Shiny module server handles the server-side logic for downloading citation metadata
}
