% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{field_codes_pubmed}
\alias{field_codes_pubmed}
\title{Field conversion codes for pubmed imports}
\format{
An object of class \code{data.frame} with 82 rows and 4 columns.
}
\usage{
field_codes_pubmed
}
\description{
Field conversion codes for pubmed imports
}
\author{
Sean Smith \email{ssmith49@ed.ac.uk}
}
\keyword{datasets}
