% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_manual.R
\name{format_cols}
\alias{format_cols}
\title{Format Manual Upload Search Data Columns for SOLES}
\usage{
format_cols(df)
}
\arguments{
\item{df}{a manual upload file dataframe to be correctly formatted for SOLES}
}
\value{
a formatted dataframe containing the required SOLES columns with standardized case
}
\description{
Internal function used to process manual uploads used to format columns so they can be combined in SOLES workflows.
}
