% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ft.R
\name{get_ft}
\alias{get_ft}
\title{Download full texts}
\usage{
get_ft(
  con,
  path,
  n = NULL,
  class_date = NULL,
  class_name = NULL,
  check_failed = FALSE,
  email = ""
)
}
\arguments{
\item{con}{connection to db - note that this must contain tables of the format 'type_tag' and dictionaries of the type type_dictionary}

\item{path}{pdf folder}

\item{n}{number of full text publications to download}

\item{class_date}{screening date to filter by}

\item{class_name}{screening name to filter by}

\item{check_failed}{logical; check failed pdfs again?}

\item{email}{email address required for UnPaywall}
}
\description{
This function downloads the full text of included citations
}
