% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_missing_metadata.R
\name{get_missing_abstracts}
\alias{get_missing_abstracts}
\title{Fill in missing abstracts via crossref}
\usage{
get_missing_abstracts(
  citations,
  abstract_col = "abstract",
  doi_col = "doi",
  id_col = "uid"
)
}
\arguments{
\item{citations}{citations you want to find abstracts for}

\item{abstract_col}{the name of the column that contains the abstract information. Default: \code{abstract}}

\item{doi_col}{the name of the column that contains the doi information. Default: \code{doi}}

\item{id_col}{a unique identifier column. For SyRF output this could be \code{CustomId}. Default: \code{uid}}
}
\value{
citations dataframe with abstracts (where feasible)
}
\description{
This function pulls in missing abstracts
}
