% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_missing_metadata.R
\name{get_missing_dois}
\alias{get_missing_dois}
\title{Fill in missing DOIs}
\usage{
get_missing_dois(citations)
}
\arguments{
\item{citations}{citations you want to find dois for}
}
\value{
citations with dois (where feasible)
}
\description{
This function pulls in missing dois from OpenAlex
}
