% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_openalex_metadata.R
\name{get_openalex_metadata}
\alias{get_openalex_metadata}
\title{Extract concepts, funder, citation count, institution and open access data from OpenAlex.}
\usage{
get_openalex_metadata(con)
}
\arguments{
\item{con}{connection to db}
}
\value{
Tables "discipline_tag", "funder_grant_tag", "institution_tag", "citation_count_tag", "retraction_tag", "article_type", oa_tag" will be updated with n number of new rows containing metadata. If no data is retrieved for these n citations, "Unknown" will be returned in the relevant columns.
}
\description{
This function retrieves meta-data.
}
\examples{
\dontrun{
# Example usage:
get_openalex_metadata(con)
}
}
