% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup.R
\name{get_recent_citations}
\alias{get_recent_citations}
\title{Pull recently retrieved citations from database}
\usage{
get_recent_citations(con, prev_months = prev_months)
}
\arguments{
\item{con}{connection to db}

\item{prev_months}{number of months to look back in database - default is 2}
}
\value{
recent citations
}
\description{
This function pulls in unique recently retrieved citations from the soles database to deduplicate
against newly retrieved citations. This ensures that any new citations are not already present in
the database (in a different format/ from a different source) before they are added to soles.
}
