% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_studies.R
\name{get_screening_decisions}
\alias{get_screening_decisions}
\title{Get Screening Decisions from the Database}
\usage{
get_screening_decisions(con, review_id = "", classifier_name = NULL)
}
\arguments{
\item{con}{A database connection object.}

\item{review_id}{A unique identifier for the review associated with the screening decisions.}

\item{classifier_name}{Name of classifier used in study_classification table (e.g. "in-vivo)}
}
\value{
A data frame containing screening decisions with columns: ITEM_ID, LABEL, TITLE, ABSTRACT, KEYWORDS, Cat, REVIEW_ID.
}
\description{
This function retrieves screening decisions from the "study_classification" table
in the specified database and returns a data frame in the correct format to run the machine learning function.
}
\examples{
\dontrun{
  screening_decisions <- get_screening_decisions(con = your_database_connection, review_id = "your_project_plus_date", classifier_name = "in-vivo")
}
}
