% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_studies.R
\name{ml_multi_validation}
\alias{ml_multi_validation}
\title{ML Multi-validation function}
\usage{
ml_multi_validation(con, review_id = "", project_name = "")
}
\arguments{
\item{con}{A database connection object.}

\item{review_id}{A string indicating the review ID.}

\item{project_name}{A string for the name of the project.}
}
\value{
Writes output to .csv files
}
\description{
This function performs a multi-validation for a machine learning algorithm, splitting data into folds
and processing it for model training, calibration, and validation.
}
\examples{
\dontrun{
ml_multi_validation(con = db_con, review_id = "my_soles_screening", project_name = "my_soles_project")
}
}
