% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{pico_multi_select_UI}
\alias{pico_multi_select_UI}
\title{PICO Multi-Select UI}
\usage{
pico_multi_select_UI(
  id,
  multi_select = TRUE,
  table,
  column,
  column2,
  label1,
  label2,
  title = "",
  selected,
  theme,
  spinner_colour
)
}
\arguments{
\item{id}{The module identifier.}

\item{multi_select}{Specifies whether to use 1 or 2 dropdown menus.}

\item{table}{The data table for generating the chart.}

\item{column}{The name of the first column for generating the chart.}

\item{column2}{The name of the second column (if multi-select is TRUE and 2 dropdown menus are required).}

\item{label1}{The label above the first input.}

\item{label2}{The label above the second input.}

\item{title}{The title of the tab.}

\item{theme}{The color status of the tab.}

\item{spinner_colour}{The color of the spinner used for loading the plot.}
}
\description{
This Shiny module creates a bar plot showing the number of publications based on the selected columns.
}
