% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_manual.R
\name{process_embase}
\alias{process_embase}
\title{Process Manually Uploaded Search results from Embase}
\usage{
process_embase(path)
}
\arguments{
\item{path}{the file path to the search results}
}
\value{
a processed dataframe containing search results
}
\description{
Internal function used to process Embase search results an format for SOLES.
}
\details{
This function reads references from the specified file using the "ovid" tag naming convention.
It performs various transformations on the data, such as uniting title and booktitle columns,
formatting pages, creating a unique identifier (uid) based on the article number (AN), and
setting the data source to "embase".
}
\examples{
\dontrun{
# Example usage:
embase_data <- process_embase("path/to/embase_data.csv")
}

}
