% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_manual.R
\name{process_medline}
\alias{process_medline}
\title{Process Manually Uploaded Search results from MEDLINE}
\usage{
process_medline(path)
}
\arguments{
\item{path}{the file path to the search results}
}
\value{
a processed dataframe containing search results
}
\description{
Internal function used to process MEDLINE search results an format for SOLES.
}
\details{
This function reads references from the specified file using the "ovid" tag naming convention.
It performs various transformations on the data, such as uniting title and booktitle columns,
formatting pages, creating a unique identifier (uid) based on the article_id, and setting
the data source to "medline".
}
\examples{
\dontrun{
# Example usage:
medline_data <- process_medline("path/to/medline_data.csv")
}

}
