% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{process_pubmed}
\alias{process_pubmed}
\title{Process PubMed references}
\usage{
process_pubmed(path)
}
\arguments{
\item{path}{The file path of the data.}
}
\value{
A processed data frame.
}
\description{
This function processes references from PubMed.
}
\details{
This function reads references from the specified file in PubMed format using the "pubmed" database source.
It adjusts column names according to the field codes, removes blank columns, adds source information,
creates a unique identifier (uid) based on the record_id, sets the data source to "pubmed", and
formats the doi column based on available information.
}
\examples{
\dontrun{
# Example usage:
pubmed_data <- process_pubmed("path/to/pubmed_data.bib")
}

}
