% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_manual.R
\name{process_pubmed}
\alias{process_pubmed}
\title{Process Manually Uploaded Search results from PubMed}
\usage{
process_pubmed(path)
}
\arguments{
\item{path}{the file path to the search results}
}
\value{
a processed dataframe containing search results
}
\description{
Internal function used to process PubMed search results an format for SOLES.
}
\details{
This function reads references from the specified file in PubMed format using the "pubmed" database source.
It adjusts column names according to the field codes, removes blank columns, adds source information,
creates a unique identifier (uid) based on the record_id, sets the data source to "pubmed", and
formats the doi column based on available information.
}
\examples{
\dontrun{
# Example usage:
pubmed_data <- process_pubmed("path/to/pubmed_data.bib")
}

}
