% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_manual.R
\name{process_scopus}
\alias{process_scopus}
\title{Process Manually Uploaded Search results from Scopus}
\usage{
process_scopus(path)
}
\arguments{
\item{path}{the file path to the search results}
}
\value{
a processed dataframe containing search results
}
\description{
Internal function used to process Scopus search results an format for SOLES.
}
\details{
This function reads references from the specified file using the "ovid" tag naming convention.
It performs various transformations on the data, such as uniting title and booktitle columns,
formatting pages, creating a unique identifier (uid) based on the Scopus EID, and setting
the data source to "scopus". Additionally, it prints the values of the article_id column.
}
\examples{
\dontrun{
# Example usage:
scopus_data <- process_scopus("path/to/scopus_data.csv")
}

}
