% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{process_wos}
\alias{process_wos}
\title{Process Web of Science references}
\usage{
process_wos(path)
}
\arguments{
\item{path}{The file path of the data.}
}
\value{
A processed data frame.
}
\description{
This function processes references from Web of Science.
}
\details{
This function reads references from the specified file in BibTeX or RIS format using the "wos" database source.
The function checks the file extension before processing accordingly.
It adjusts column names according to the field codes, removes blank columns, adds source information,
creates a unique identifier (uid) by modifying the existing uid column, and sets the data source to "wos".
}
\examples{
\dontrun{
# Example usage:
wos_data <- process_wos("path/to/wos_data.bib")
}

}
