% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_api.R
\name{pubmed_search}
\alias{pubmed_search}
\title{Search PubMed and retrieve bibliographic data using RISmed}
\usage{
pubmed_search(query, timespan, retMax = 5000, format_soles = TRUE)
}
\arguments{
\item{query}{a character string containing a correctly syntaxed PubMed search}

\item{timespan}{a formatted character string defining the timespan you want to search}

\item{retMax}{The maximum number of records to retrieve, default is 5000, maximum is 5000}

\item{format_soles}{boolean, if set to TRUE will format search results for the SOLES workflow, default is TRUE}
}
\value{
a dataframe containing Web of Science search results
}
\description{
A wrapper function for RISmed (https://github.com/kaitlynhair/RISmed). Search PubMed using a query and retrieve results programmatically.
The timespan should be formatted as a number followed by the word "week" or "month", e.g. "1month" or "2week".
One week is calculated as 7 days and one month is calculated as 31 days.
}
\examples{
\dontrun{
query <- "(dementia[tiab] OR \"memory loss\"[tiab])"
pubmed_result <- pubmed_search(query, timespan = "1week", format_soles = FALSE)
}
}
