% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_studies.R
\name{run_error_correction}
\alias{run_error_correction}
\title{Run Error Correction}
\usage{
run_error_correction(
  con,
  k_fold_scores = as.character(),
  k_fold_thresholds = as.character(),
  type = "extreme discrepancies",
  number_to_be_re_reviewed = as.numeric()
)
}
\arguments{
\item{con}{A database connection object.}

\item{k_fold_scores}{A character string specifying the file path to the CSV containing machine scores.}

\item{k_fold_thresholds}{A character string specifying the file path to the CSV containing performance metrics.}

\item{type}{A character string specifying the type of selection method. Options are "extreme discrepancies" or "random".}

\item{number_to_be_re_reviewed}{An integer specifying the number of studies to be re-reviewed.}
}
\value{
A dataframe containing the selected studies for re-review.
}
\description{
This function identifies and selects studies for re-review based on disagreements between human and machine decisions.
This function takes the results from the run_k_fold function on 1 repeat, which will contain a human screening decision and a machine score for each study.
}
\examples{
\dontrun{
# Run error correction with default parameters
run_error_correction(con = my_connection,
                     k_fold_scores = "k_fold_scores.csv",
                     k_fold_thresholds = "k_fold_performance.csv",
                     type = "extreme discrepancies",
                     number_to_be_re_reviewed = 100)
}

}
