% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_studies.R
\name{run_k_fold}
\alias{run_k_fold}
\title{Run K-Fold Cross-Validation for Machine Learning}
\usage{
run_k_fold(
  con,
  project_name = as.character(),
  classifier_name = as.character(),
  screening_decisions_review_id = as.character(),
  fold_number = 5
)
}
\arguments{
\item{con}{A database connection object.}

\item{project_name}{A string specifying the project name. Used for naming output files and directories.}

\item{classifier_name}{A string representing the classifier's name.}

\item{screening_decisions_review_id}{A string indicating the review ID for retrieving human screening decisions.}

\item{fold_number}{An integer specifying the number of folds for cross-validation. Default is 5.}
}
\value{
The function does not return an R object. It writes the results to the \code{k-fold-validation} directory.
}
\description{
This function performs K-fold cross-validation. It retrieves data
from a database, splits it into \code{k} folds, trains and tests a classifier for each fold, and
computes performance metrics. Results are saved to CSV files for further analysis.
}
\details{
This function:
\enumerate{
\item Retrieves included and excluded screening decisions from the database.
\item Shuffles the data and splits it into the specified number of folds.
\item For each iteration:
\itemize{
\item Assigns a test set while others are used for training.
\item Runs the specified machine learning classifier via an external API.
\item Computes and saves performance metrics and test scores.
}
}

Results are saved in a directory named \code{k-fold-validation} in the current working directory.
}
\examples{
\dontrun{
# Example usage:
run_k_fold(
  con = db_con,
  project_name = "MS-SOLES",
  classifier_name = "in-vivo",
  screening_decisions_review_id = "ms-soles-screening",
  fold_number = 5
)
}
}
