% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{scopus_search}
\alias{scopus_search}
\title{Get Scopus results}
\usage{
scopus_search(query, api_key, retMax = 2000)
}
\arguments{
\item{query}{search query}

\item{api_key}{scopus api key}

\item{retMax}{maximum number of citations retrieved, specify based on timeframe}
}
\value{
dataframe of search results
}
\description{
This function retrieves search results from Scopus
}
