% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{search_Server}
\alias{search_Server}
\title{Search Page Module}
\usage{
search_Server(
  id,
  table,
  combined_pico_table,
  pico_data = list(),
  citations_for_download,
  project_name = ""
)
}
\arguments{
\item{id}{The module identifier.}

\item{table}{A dataframe of included studies with metadata.}

\item{combined_pico_table}{A combined dataframe of PICO tags.}

\item{pico_data}{A list of dynamic search updates based on the PICO dropdown filters.}

\item{citations_for_download}{A dataframe containing citations for download.}

\item{project_name}{The name of the project.}
}
\description{
This Shiny module generates the search page for the app, allowing users to search and filter studies.
}
