% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_soles}
\alias{setup_soles}
\title{This function reads in search files in the soles project directory and populates the \code{unique_citations} and \verb{retrieved_citations tables}.
\code{setup_soles()} expects that the required tables in the project database already exist. See \code{create_tables()}.}
\usage{
setup_soles(con, folder, master_node)
}
\arguments{
\item{con}{Connection to db}

\item{folder}{The name of the folder in the project directory where all search export files are saved.}

\item{master_node}{OSF master node}
}
\description{
This function reads in search files in the soles project directory and populates the \code{unique_citations} and \verb{retrieved_citations tables}.
\code{setup_soles()} expects that the required tables in the project database already exist. See \code{create_tables()}.
}
\details{
The function expects that within the main search folder, each sub-folder is named after the corresponding database source.
Valid sub-folder names are: pubmed, medline, wos, scopus, ovid, psychinfo, embase.
The function also expects that each search file within the sub-folder follows the consistent naming convention \code{source_date}, where \code{source}
is the literature database and \code{date} corresponds to the date the search was conducted in \code{ddmmyyyy} format.
For example, the following would be valid search export file names: \code{pubmed_01012023.txt}, \code{scopus_15102022.bib}.
If multiple export files from the same search on the same date exist, these can be distinguished by numbering the searches
e.g. \code{pubmed1_01012023.txt}, \code{pubmed2_01012023.txt}, etc. Full path example: \code{manual_search/pubmed/pubmed1_01012023.txt}
}
\examples{
\dontrun{
# Example usage:
setup_soles(con, "path/to/search/folder")
setup_soles(con, "manual_search")
}

}
