% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_api.R
\name{wos_search}
\alias{wos_search}
\title{Search Web of Science Core Collection and retrieve bibliographic data using rwoslite}
\usage{
wos_search(query = NULL, timespan = NULL, format_soles = TRUE)
}
\arguments{
\item{query}{a character string containing a correctly syntaxed Web of Science search}

\item{timespan}{a formatted character string defining the timespan you want to search}

\item{format_soles}{boolean, if set to TRUE will format search results for the SOLES workflow, default is TRUE}
}
\value{
a dataframe containing Web of Science search results
}
\description{
A wrapper function for rwoslite / WOS Lite. Search the Web of Science Core Collection using a query and retrieve results programmatically.
Requires an API key from the Clarivate Developer Portal (https://developer.clarivate.com/apis/woslite).
Use the function \code{usethis::edit_r_environ()} to add the key to your ~/.Renviron file.
The timespan should be formatted as a number followed by the word "week" or "month", e.g. "1month" or "2week".
One week is calculated as 7 days and one month is calculated as 31 days.
}
\examples{
\dontrun{
# api_key should be stored in your r environment and not shared
query <- "TS=(dementia OR \"memory loss\")"
wos_result <- wos_search(query, timespan = "1week", format_soles = FALSE)
}
}
