% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{wos_search}
\alias{wos_search}
\title{Get Web of Science results
This function retrieves search results from pubmed	#' This function retrieves search results from Web of Science, using the WOS Lite API. A personal API key
must be requested from the Clarivate developer portal (https://developer.clarivate.com/apis/woslite)
and stored as an R Environment variable.}
\usage{
wos_search(query, timespan)
}
\arguments{
\item{query}{search query}

\item{timespan}{time frame for search, options are: 1week, 2week, 1month}
}
\value{
dataframe of search results
}
\description{
Use the function \code{usethis::edit_r_environ()} to add the key to your ~/.Renviron file:
}
