% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{yearBarUI}
\alias{yearBarUI}
\title{Generate a Yearly Bar Chart}
\usage{
yearBarUI(id, title = "", theme = "", spinner_colour = "#96c296", table)
}
\arguments{
\item{id}{The module identifier.}

\item{title}{The title of the tab.}

\item{theme}{The color (status) of the tab.}

\item{spinner_colour}{The color of the spinner shown during plot generation.}

\item{table}{The input data table.}
}
\description{
This Shiny module creates a bar chart displaying the count of occurrences by year for a selected column in a given table. The chart includes a legend with a color-coded representation of the column values. This module is designed to be used within a Shiny application.
}
